TEZ-4137: Input/Output/Processor should merge payload to local conf
Runtime components do not have access to session config. Because of that, we have to send all config options to every task eventhough a large part of the configs are in session config, just not accessible to the task. This patch makes it accessible to the tasks. All input/output/processor s can merge their user payload config to this session config. Therefore, we can send less bytes over the wire, saving time in serializing and transmitting configs client-to-am and am-to-task communication.

Change-Id: I8cc0d17deec1e03793b884fd05460e3e69afab6a