TEZ-4137: Merge runtime component configs with session config
Runtime components do not have access to session config. Because of that, we have to send all config options to every task eventhough a large part of the configs are in session config, just not accessible to the task. This patch makes it accessible to the tasks. All input/output/processor s can merge their user payload config to this session config. Therefore, we can send less bytes over the wire, saving time in serializing and transmitting configs client-to-am and am-to-task communication.

Change-Id: I995ed36ae0f5a71e12783802237b768d4fda56ca